packageの編集に合わせてcomponentsを編集。getshapeをGenerateModelFile.pyへ移動
