+++ packages/GeneratePreprocessFile.py
@@ -76,13 +76,7 @@ class PreprocessInfo:
             self.Prep_image_dataset(dicts, 'test', dataset_path, project_path)
         else:
             return
-    
-    def get_shape(image_size=(None, None), color_mode='rgb'):    
-        if color_mode == 'rgb':
-            color = 3
-        else:
-            color = 1
-        return (image_size[0], image_size[1], color)
+
 
 if __name__ == '__main__':
     preprocess_dicts = {
@@ -99,8 +93,6 @@ if __name__ == '__main__':
     prep = PreprocessInfo(data_type='image')
     prep.Prep_dataset(preprocess_dicts)
 
-    shape_size = PreprocessInfo.get_shape(image_size=(256, 256), color_mode='rgb')
-    print(shape_size)
